+++ gakumas-tools/components/EntityBank/EntityBank.js
@@ -15,7 +15,7 @@ import {
 import styles from "./EntityBank.module.scss";
 
 const HIDDEN_ITEM_IDS = [419];
-const HIDDEN_CARD_IDS = [803];
+const HIDDEN_CARD_IDS = [803, 806];
 
 function EntityBank({ type, onClick, filters = [], includeNull = true }) {
   const t = useTranslations("EntityBank");