+++ packages/gakumas-images/images/pItems/icons/imports.js
@@ -417,6 +417,7 @@ import image_416 from "./416.png";
 import image_417 from "./417.png";
 import image_418 from "./418.png";
 import image_420 from "./420.png";
+import image_421 from "./421.png";
 
 const IMAGES = {
   '1': image_1,
@@ -838,6 +839,7 @@ const IMAGES = {
   '417': image_417,
   '418': image_418,
   '420': image_420,
+  '421': image_421,
 };
 
 export default IMAGES;